requirements.txt
fastapi==0.115.6
uvicorn==0.32.1

python-dotenv==1.0.1
requests==2.32.3

openai==1.59.3

pydantic==2.10.4
pydantic-settings==2.7.0

sqlalchemy==2.0.36
psycopg2-binary==2.9.9

upstash-redis==1.2.0

python-multipart==0.0.9


sqlalchemy>=2.0
psycopg2-binary


#twilio

httpx
